Apply edges by priority
Apply edges in the compartment graph in ascending order of
their `priority` attributes. Edges with negative or null
`priority`, and edges without the attr, are applied last
by default. Edges that have the same priority will have some
scaling factor k applied to their weights before applying,
to ensure that the population of the origin node does not
decrease below 0.